requirements/base.txt
redis==5.0.*
alembic==1.11.*
bcrypt==4.0.*
psycopg2==2.9.*
python-jose==3.3.*
SQLAlchemy==2.0.20
httpx==0.24.*

pydantic[email]==2.3.*
pydantic-settings==2.0.*
asyncpg==0.28.*
fastapi==0.103.*
uvicorn[standard]==0.23.*
pytest-cov

sentry-sdk==1.29.*
